Promote Service EP preloading to 'error' level violation
