+++ intellij-plugin-structure/structure-intellij/src/main/java/com/jetbrains/plugin/structure/intellij/problems/InvalidDescriptorErrors.kt
@@ -6,6 +6,7 @@ package com.jetbrains.plugin.structure.intellij.problems
 
 import com.jetbrains.plugin.structure.base.plugin.PluginProblem
 import com.jetbrains.plugin.structure.base.problems.InvalidDescriptorProblem
+import com.jetbrains.plugin.structure.intellij.plugin.IdePluginContentDescriptor
 import com.jetbrains.plugin.structure.intellij.version.IdeVersion
 
 class PropertyWithDefaultValue(
@@ -199,4 +200,20 @@ class OptionalDependencyConfigFileIsEmpty(private val optionalDependencyId: Stri
 
   override val detailedMessage: String
     get() = "Optional dependency declaration on '$optionalDependencyId' cannot have empty \"config-file\""
+}
+
+class ServiceExtensionPointPreloadNotSupported(private val serviceType: IdePluginContentDescriptor.ServiceType) : PluginProblem() {
+  private val extensionPointPrefix = "com.intellij"
+
+  override val level
+    get() = Level.ERROR
+
+  override val message
+    get() = "Service preloading is deprecated in the <${serviceType.toXmlElement()}> element. Consider removing the 'preload' attribute and migrating to the listeners, see https://plugins.jetbrains.com/docs/intellij/plugin-listeners.html"
+
+  private fun IdePluginContentDescriptor.ServiceType.toXmlElement(): String = "$extensionPointPrefix." + when (this) {
+    IdePluginContentDescriptor.ServiceType.PROJECT -> "projectService"
+    IdePluginContentDescriptor.ServiceType.APPLICATION -> "applicationService"
+    IdePluginContentDescriptor.ServiceType.MODULE -> "moduleService"
+  }
 }
